+++ pure_protobuf/__init__.py
@@ -3,11 +3,7 @@
 
 """
 Python implementation of Protocol Buffers data types.
-
-`pure-protobuf` contributors © 2011-2019
 """
 
-import sys
-
-if sys.version_info >= (3, 6, 0, 0):
-    from pure_protobuf.serializers import read_varint, write_varint  # noqa: F401
+import pure_protobuf.serializers.google  # noqa: F401
+from pure_protobuf.serializers import read_varint, write_varint  # noqa: F401